feat(V5): adapt remaining C++ binaries + fix fp16 dtype mismatch (Phase 5+6)
Phase 5: gomoku_play_main, gomoku_eval_main, mcts_probe_main all updated
to V5 inference path (encode_state_v5, double-input forward({state, global}),
dict output unpack with policy.select(1, 0) for main head, RuleType
constructor). gomoku_eval_main's BatchedInferenceServer takes optional
Gomoku* for global feature derivation.

Phase 6: end-to-end pipeline now functional under V5.

Bug fixed during validation: KataGPool / KataValueHeadGPool were forcing
float32 mid-stream (`dtype=torch.float32` in sum reduction, `.to(torch.float32)`
on max), but the C++ inference path casts the entire model to half (kHalf) on
CUDA. The downstream `linear_g` weight was fp16 while gpool output was fp32 →
matmul dtype mismatch. Fixed by removing the fp32 forcing; gpool now preserves
input dtype throughout. KataGoModel originally used fp32 throughout the model;
SkyZero's mixed-half mode requires consistent dtype.

End-to-end validation:
1. cmake build: all 4 binaries compile (selfplay_main, gomoku_eval, gomoku_play, mcts_probe)
2. python/init_model.py: produces valid TorchScript latest.pt (b8c96 v2)
3. selfplay_main: cold-start 1 game → NPZ with 7 keys, state (3,5,15,15) int8,
   global_features (3,12) float32 with correct rule one-hot + renju_color_sign=-1
   + has_forbidden=1 + ply/225 normalized
4. python/train.py: 20 steps over the NPZ → loss monotonically decreases
   (total 7.32→6.15, value 1.08→0.25), no NaN

71 unit tests still passing after gpool dtype change.

Phase 4b (per-game RULE_DISTRIBUTION + BOARD_SIZE_DISTRIBUTION) still TODO;
current setup is single-rule single-15×15 selected via run.cfg RULE=renju.

+++ SkyZero_V5/python/nets_v2.py
@@ -164,8 +164,11 @@ class KataGPool(nn.Module):
         if mask_sum_hw is None:
             mask_sum_hw = mask.sum(dim=(2, 3), keepdim=True)
         sqrt_off = torch.sqrt(mask_sum_hw) - 14.0
-        layer_mean = torch.sum(x * mask, dim=(2, 3), keepdim=True, dtype=torch.float32) / mask_sum_hw
-        layer_max = (x + (mask - 1.0)).to(torch.float32).amax(dim=(2, 3), keepdim=True)
+        # Preserve input dtype for fp16 inference compatibility (downstream linear_g
+        # weight is in input dtype). KataGoModel used fp32 here for precision under
+        # an fp32-only model; SkyZero's selfplay casts the whole module to half.
+        layer_mean = torch.sum(x * mask, dim=(2, 3), keepdim=True) / mask_sum_hw
+        layer_max = (x + (mask - 1.0)).amax(dim=(2, 3), keepdim=True)
         return torch.cat((
             layer_mean,
             layer_mean * (sqrt_off / 10.0),
@@ -181,7 +184,8 @@ class KataValueHeadGPool(nn.Module):
         if mask_sum_hw is None:
             mask_sum_hw = mask.sum(dim=(2, 3), keepdim=True)
         sqrt_off = torch.sqrt(mask_sum_hw) - 14.0
-        layer_mean = torch.sum(x * mask, dim=(2, 3), keepdim=True, dtype=torch.float32) / mask_sum_hw
+        # Preserve input dtype (see KataGPool note above re: fp16 compat).
+        layer_mean = torch.sum(x * mask, dim=(2, 3), keepdim=True) / mask_sum_hw
         return torch.cat((
             layer_mean,
             layer_mean * (sqrt_off / 10.0),
